Pushing user-specific changes
Figured out how to work with the test framework
Changed filter system to use a string instead of a boolean to allow for future expansion

UI works as expected, but tests fail.  I don't know the proper way to handle the checkboxes to work with the testing framework.

I'm not a React expert, I need to move on to the other questions

Adding Table of People
Adding Sorts
Implementing sorts in filters